Refactor css transaction components
React Hooks shouldn't be used in loops or if statement.
