+++ client/app/bundles/comments/src/CSSTransition/CSSTransition.bs.mjs
@@ -1,0 +1,12 @@
+// Generated by ReScript, PLEASE EDIT WITH CARE
+
+
+var TransitionGroup = {};
+
+var CSSTransition = {};
+
+export {
+  TransitionGroup ,
+  CSSTransition ,
+}
+/* No side effect */
