+++ client/app/bundles/comments/src/CommentList/AlertError/AlertError.bs.mjs
@@ -1,46 +1,29 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as React from "react";
-import * as JsxRuntime from "react/jsx-runtime";
-import * as ReactTransitionGroup from "react-transition-group";
+import * as React from 'react';
+import * as JsxRuntime from 'react/jsx-runtime';
+import * as ReactTransitionGroup from 'react-transition-group';
 
 function AlertError(props) {
-  var cssTransitionGroupClassNames = props.cssTransitionGroupClassNames;
   var nodeRef = React.useRef(null);
-  switch (props.error) {
-    case /* NoError */0 :
-        return null;
-    case /* FailedToSaveComment */1 :
-        return JsxRuntime.jsx(ReactTransitionGroup.CSSTransition, {
-                    children: JsxRuntime.jsx("div", {
-                          children: JsxRuntime.jsx("strong", {
-                                children: "Your comment was not saved!"
-                              }),
-                          className: "alert alert-danger"
-                        }),
-                    timeout: 500,
-                    nodeRef: nodeRef,
-                    classNames: cssTransitionGroupClassNames
-                  }, "commentFetchError");
-    case /* FailedToFetchComments */2 :
-        return JsxRuntime.jsx(ReactTransitionGroup.CSSTransition, {
-                    children: JsxRuntime.jsx("div", {
-                          children: JsxRuntime.jsx("strong", {
-                                children: "Can't fetch the comments!"
-                              }),
-                          className: "alert alert-danger"
-                        }),
-                    timeout: 500,
-                    nodeRef: nodeRef,
-                    classNames: cssTransitionGroupClassNames
-                  }, "commentFetchError");
-    
-  }
+  return JsxRuntime.jsx(
+    ReactTransitionGroup.CSSTransition,
+    {
+      children: JsxRuntime.jsx('div', {
+        children: JsxRuntime.jsx('strong', {
+          children: 'Your comment was not saved!',
+        }),
+        className: 'alert alert-danger',
+      }),
+      timeout: 500,
+      nodeRef: nodeRef,
+      classNames: props.cssTransitionGroupClassNames,
+    },
+    'commentFetchError',
+  );
 }
 
 var make = AlertError;
 
-export {
-  make ,
-}
+export { make };
 /* react Not a pure module */
