+++ client/app/bundles/comments/src/CommentList/Comment/Comment.bs.mjs
@@ -1,9 +1,11 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
+import * as React from "react";
 import * as Marked from "marked";
 import * as Caml_option from "rescript/lib/es6/caml_option.js";
 import * as JsxRuntime from "react/jsx-runtime";
 import CommentModuleScss from "./Comment.module.scss";
+import * as ReactTransitionGroup from "react-transition-group";
 
 var css = CommentModuleScss;
 
@@ -15,19 +17,25 @@ function $$Comment(props) {
   var innerHTML = {
     __html: rawMarkup
   };
-  return JsxRuntime.jsxs("div", {
-              children: [
-                JsxRuntime.jsx("h2", {
-                      children: comment.author,
-                      className: css.commentAuthor
-                    }),
-                JsxRuntime.jsx("span", {
-                      dangerouslySetInnerHTML: Caml_option.some(innerHTML)
-                    })
-              ],
-              ref: Caml_option.some(props.nodeRef),
-              className: css.comment
-            });
+  var nodeRef = React.useRef(null);
+  return JsxRuntime.jsx(ReactTransitionGroup.CSSTransition, {
+              children: JsxRuntime.jsxs("div", {
+                    children: [
+                      JsxRuntime.jsx("h2", {
+                            children: comment.author,
+                            className: css.commentAuthor
+                          }),
+                      JsxRuntime.jsx("span", {
+                            dangerouslySetInnerHTML: Caml_option.some(innerHTML)
+                          })
+                    ],
+                    ref: Caml_option.some(nodeRef),
+                    className: css.comment
+                  }),
+              timeout: 500,
+              nodeRef: nodeRef,
+              classNames: props.cssTransitionGroupClassNames
+            }, "component_" + String(comment.id));
 }
 
 var make = $$Comment;
