+++ client/app/bundles/comments/src/CommentList/CommentList.bs.mjs
@@ -1,17 +1,13 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as React from "react";
 import * as $$Comment from "./Comment/Comment.bs.mjs";
+import * as AlertError from "./AlertError/AlertError.bs.mjs";
 import * as Belt_Array from "rescript/lib/es6/belt_Array.js";
 import * as JsxRuntime from "react/jsx-runtime";
 import * as ReactTransitionGroup from "react-transition-group";
 import RescriptPageModuleScss from "../RescriptPage.module.scss";
 
 var css = RescriptPageModuleScss;
-
-var TransitionGroup = {};
-
-var CSSTransition = {};
 
 function CommentList(props) {
   var cssTransitionGroupClassNames_enter = css.elementEnter;
@@ -24,34 +20,14 @@ function CommentList(props) {
     exit: cssTransitionGroupClassNames_exit,
     exitActive: cssTransitionGroupClassNames_exitActive
   };
-  var errorMsg;
-  if (props.error) {
-    var nodeRef = React.useRef(null);
-    errorMsg = JsxRuntime.jsx(ReactTransitionGroup.CSSTransition, {
-          children: JsxRuntime.jsx("div", {
-                children: JsxRuntime.jsx("strong", {
-                      children: "Your comment was not saved!"
-                    }),
-                className: "alert alert-danger"
-              }),
-          timeout: 500,
-          nodeRef: nodeRef,
-          classNames: cssTransitionGroupClassNames
-        }, "commentFetchError");
-  } else {
-    errorMsg = JsxRuntime.jsx(JsxRuntime.Fragment, {});
-  }
+  var errorMsg = props.error ? JsxRuntime.jsx(AlertError.make, {
+          cssTransitionGroupClassNames: cssTransitionGroupClassNames
+        }) : JsxRuntime.jsx(JsxRuntime.Fragment, {});
   var commentsElements = Belt_Array.map(props.comments, (function (comment) {
-          var nodeRef = React.useRef(null);
-          return JsxRuntime.jsx(ReactTransitionGroup.CSSTransition, {
-                      children: JsxRuntime.jsx($$Comment.make, {
-                            comment: comment,
-                            nodeRef: nodeRef
-                          }),
-                      timeout: 500,
-                      nodeRef: nodeRef,
-                      classNames: cssTransitionGroupClassNames
-                    }, "component_" + String(comment.id));
+          return JsxRuntime.jsx($$Comment.make, {
+                      comment: comment,
+                      cssTransitionGroupClassNames: cssTransitionGroupClassNames
+                    }, "comment_" + String(comment.id));
         }));
   return JsxRuntime.jsxs("div", {
               children: [
@@ -69,8 +45,6 @@ var make = CommentList;
 
 export {
   css ,
-  TransitionGroup ,
-  CSSTransition ,
   make ,
 }
 /* css Not a pure module */
