+++ client/app/bundles/comments/src/CommentList/CommentList.res
@@ -1,35 +1,8 @@
 @module("../RescriptPage.module.scss") external css: {..} = "default"
 
-module TransitionGroup = {
-  @react.component @module("react-transition-group")
-  external make: (
-    ~children: React.element,
-    ~className: string,
-    ~component: string
-  ) => React.element = "TransitionGroup"
-}
-
-type cssTransitionGroupClassNamesT = {
-  enter: string,
-  enterActive: string,
-  exit: string,
-  exitActive: string
-}
-
-module CSSTransition = {
-  @react.component @module("react-transition-group")
-  external make: (
-    ~children: React.element, 
-    ~key: string, 
-    ~timeout: int, 
-    ~nodeRef: React.ref<Js.Nullable.t<'a>>, 
-    ~classNames: cssTransitionGroupClassNamesT
-  ) => React.element = "CSSTransition"
-}
-
 @react.component
 let make = (~comments: Actions.Api.comments, ~error: bool) => {
-  let cssTransitionGroupClassNames = {
+  let cssTransitionGroupClassNames: CSSTransition.cssTransitionGroupClassNamesT = {
     enter: css["elementEnter"],
     enterActive: css["elementEnterActive"],
     exit: css["elementLeave"],
@@ -37,46 +10,18 @@ let make = (~comments: Actions.Api.comments, ~error: bool) => {
   }
 
   let errorMsg = switch error {
-  | true => (() => {
-      let nodeRef = React.useRef(Js.Nullable.null)
-      
-      // The 500 must correspond to the 0.5s in:
-      //   ../RescriptPage.module.scss:8
-      <CSSTransition
-        key="commentFetchError"
-        nodeRef={nodeRef}
-        timeout={500}
-        classNames={cssTransitionGroupClassNames}
-      >
-        <div className="alert alert-danger">
-          <strong>{React.string("Your comment was not saved!")}</strong>
-        </div>
-      </CSSTransition>
-    })()
+  | true => <AlertError cssTransitionGroupClassNames=cssTransitionGroupClassNames />
   | false => <></>
   }
 
   let commentsElements = Belt.Array.map(
     comments,
-    comment => {
-      let nodeRef = React.useRef(Js.Nullable.null)
-      
-      // The 500 must correspond to the 0.5s in:
-      //   ../RescriptPage.module.scss:8
-      <CSSTransition
-        key={"component_" ++ Belt.Int.toString(comment.id) }
-        timeout={500}
-        nodeRef={nodeRef}
-        classNames=cssTransitionGroupClassNames
-      >
-        <Comment comment=comment nodeRef=nodeRef/>
-      </CSSTransition>
-    });
+    comment => <Comment comment=comment cssTransitionGroupClassNames key={"comment_" ++ Belt.Int.toString(comment.id)} />);
 
   <div>
     errorMsg
-    <TransitionGroup className="commentList" component="div">
+    <CSSTransition.TransitionGroup className="commentList" component="div">
       {React.array(commentsElements)}
-    </TransitionGroup>
+    </CSSTransition.TransitionGroup>
   </div>
 }
