+++ client/app/bundles/comments/src/RescriptPage.module.scss
@@ -1,5 +1,6 @@
 // The 0.5s must correspond to the 500s in:
-//   ./CommentList/CommentList.res:47
+//   ./CommentList/AlertError/AlertError.res:10
+//   ./CommentList/Comment/Comment.res:18
 .elementEnter {
   opacity: 0.01;
